--- src/App.css
@@ -1,5 +1,0 @@
-.App {
-  width: 100%;
-  max-width: 1136px;
-  margin: 10px auto 0;
-}
