+++ src/components/App/styles.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const StyledAppContainer = styled.div`
+  width: 100%;
+  max-width: 1136px;
+  margin: 10px auto 0;
+`;
