+++ src/components/TemplateTypes/index.tsx
@@ -1,4 +1,8 @@
-import React, { FC, useEffect, useState } from "react";
+import React, {
+  FC,
+  useEffect,
+  useState,
+} from "react";
 import { ITemplateTypes } from "./types";
 import * as SC from "./styles";
