+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
-import App from "./App";
+import App from "./components/App";
 
 const root = createRoot(document.getElementById("root") as HTMLElement);
 root.render(<App />);
